// frontend/src/app/pages/chats-page/libs/openai/client.d.ts
export class RealtimeClient extends RealtimeEventHandler {
  /**
   * Create a new RealtimeClient instance
   * @param {{url?: string, apiKey?: string, dangerouslyAllowAPIKeyInBrowser?: boolean, debug?: boolean}} [settings]
   */
  constructor({
    url,
    apiKey,
    dangerouslyAllowAPIKeyInBrowser,
    debug,
  }?: {
    url?: string;
    apiKey?: string;
    dangerouslyAllowAPIKeyInBrowser?: boolean;
    debug?: boolean;
  });
  defaultSessionConfig: {
    modalities: string[];
    instructions: string;
    voice: string;
    input_audio_format: string;
    output_audio_format: string;
    input_audio_transcription: any;
    turn_detection: any;
    tools: any[];
    tool_choice: string;
    temperature: number;
    max_response_output_tokens: number;
  };
  sessionConfig: {};
  transcriptionModels: {
    model: string;
  }[];
  defaultServerVadConfig: {
    type: string;
    threshold: number;
    prefix_padding_ms: number;
    silence_duration_ms: number;
  };
  realtime: RealtimeAPI;
  conversation: RealtimeConversation;
  /**
   * Resets sessionConfig and conversationConfig to default
   * @private
   * @returns {true}
   */
  private _resetConfig;
  sessionCreated: boolean;
  tools: {};
  inputAudioBuffer: any;
  /**
   * Sets up event handlers for a fully-functional application control flow
   * @private
   * @returns {true}
   */
  private _addAPIEventHandlers;
  /**
   * Tells us whether the realtime socket is connected and the session has started
   * @returns {boolean}
   */
  isConnected(): boolean;
  /**
   * Resets the client instance entirely: disconnects and clears active config
   * @returns {true}
   */
  reset(): true;
  /**
   * Connects to the Realtime WebSocket API
   * Updates session config and conversation config
   * @returns {Promise<true>}
   */
  connect(): Promise<true>;
  /**
   * Waits for a session.created event to be executed before proceeding
   * @returns {Promise<true>}
   */
  waitForSessionCreated(): Promise<true>;
  /**
   * Disconnects from the Realtime API and clears the conversation history
   */
  disconnect(): void;
  /**
   * Gets the active turn detection mode
   * @returns {"server_vad"|null}
   */
  getTurnDetectionType(): 'server_vad' | null;
  /**
   * Add a tool and handler
   * @param {ToolDefinitionType} definition
   * @param {function} handler
   * @returns {{definition: ToolDefinitionType, handler: function}}
   */
  addTool(
    definition: ToolDefinitionType,
    handler: Function
  ): {
    definition: ToolDefinitionType;
    handler: Function;
  };
  /**
   * Removes a tool
   * @param {string} name
   * @returns {true}
   */
  removeTool(name: string): true;
  /**
   * Deletes an item
   * @param {string} id
   * @returns {true}
   */
  deleteItem(id: string): true;
  /**
   * Updates session configuration
   * If the client is not yet connected, will save details and instantiate upon connection
   * @param {SessionResourceType} [sessionConfig]
   */
  updateSession({
    modalities,
    instructions,
    voice,
    input_audio_format,
    output_audio_format,
    input_audio_transcription,
    turn_detection,
    tools,
    tool_choice,
    temperature,
    max_response_output_tokens,
  }?: SessionResourceType): boolean;
  /**
   * Sends user message content and generates a response
   * @param {Array<InputTextContentType|InputAudioContentType>} content
   * @returns {true}
   */
  sendUserMessageContent(
    content?: Array<InputTextContentType | InputAudioContentType>
  ): true;
  /**
   * Appends user audio to the existing audio buffer
   * @param {Int16Array|ArrayBuffer} arrayBuffer
   * @returns {true}
   */
  appendInputAudio(arrayBuffer: Int16Array | ArrayBuffer): true;
  /**
   * Forces a model response generation
   * @returns {true}
   */
  createResponse(): true;
  /**
   * Cancels the ongoing server generation and truncates ongoing generation, if applicable
   * If no id provided, will simply call `cancel_generation` command
   * @param {string} id The id of the message to cancel
   * @param {number} [sampleCount] The number of samples to truncate past for the ongoing generation
   * @returns {{item: (AssistantItemType | null)}}
   */
  cancelResponse(
    id: string,
    sampleCount?: number
  ): {
    item: AssistantItemType | null;
  };
  /**
   * Utility for waiting for the next `conversation.item.appended` event to be triggered by the server
   * @returns {Promise<{item: ItemType}>}
   */
  waitForNextItem(): Promise<{
    item: ItemType;
  }>;
  /**
   * Utility for waiting for the next `conversation.item.completed` event to be triggered by the server
   * @returns {Promise<{item: ItemType}>}
   */
  waitForNextCompletedItem(): Promise<{
    item: ItemType;
  }>;
}
/**
 * Valid audio formats
 */
export type AudioFormatType = 'pcm16' | 'g711_ulaw' | 'g711_alaw';
export type AudioTranscriptionType = {
  model: 'whisper-1';
};
export type TurnDetectionServerVadType = {
  type: 'server_vad';
  threshold?: number;
  prefix_padding_ms?: number;
  silence_duration_ms?: number;
};
/**
 * Tool definitions
 */
export type ToolDefinitionType = {
  type?: 'function';
  name: string;
  description: string;
  parameters: {
    [key: string]: any;
  };
};
export type SessionResourceType = {
  model?: string;
  modalities?: string[];
  instructions?: string;
  voice?:
    | 'alloy'
    | 'ash'
    | 'ballad'
    | 'coral'
    | 'echo'
    | 'sage'
    | 'shimmer'
    | 'verse';

  input_audio_format?: AudioFormatType;
  output_audio_format?: AudioFormatType;
  input_audio_transcription?: AudioTranscriptionType | null;
  turn_detection?: TurnDetectionServerVadType | null;
  tools?: ToolDefinitionType[];
  tool_choice?:
    | 'auto'
    | 'none'
    | 'required'
    | {
        type: 'function';
        name: string;
      };
  temperature?: number;
  max_response_output_tokens?: number | 'inf';
};
export type ItemStatusType = 'in_progress' | 'completed' | 'incomplete';
export type InputTextContentType = {
  type: 'input_text';
  text: string;
};
export type InputAudioContentType = {
  type: 'input_audio';
  /**
   * base64-encoded audio data
   */
  audio?: string;
  transcript?: string | null;
};
export type TextContentType = {
  type: 'text';
  text: string;
};
export type AudioContentType = {
  type: 'audio';
  /**
   * base64-encoded audio data
   */
  audio?: string;
  transcript?: string | null;
};
export type SystemItemType = {
  previous_item_id?: string | null;
  type: 'message';
  status: ItemStatusType;
  role: 'system';
  content: Array<InputTextContentType>;
};
export type UserItemType = {
  previous_item_id?: string | null;
  type: 'message';
  status: ItemStatusType;
  role: 'user';
  content: Array<InputTextContentType | InputAudioContentType>;
};
export type AssistantItemType = {
  previous_item_id?: string | null;
  type: 'message';
  status: ItemStatusType;
  role: 'assistant';
  content: Array<TextContentType | AudioContentType>;
};
// ...
export type FormattedItemType = {
  id: string;
  object: string;
  role?: 'user' | 'assistant' | 'system';
  formatted: FormattedPropertyType;
};
export type BaseItemType =
  | SystemItemType
  | UserItemType
  | AssistantItemType
  | FunctionCallItemType
  | FunctionCallOutputItemType;
export type ItemType = FormattedItemType & BaseItemType;